Correct the free-list safety argument and harden the new tests
Adversarial review found the justification for dropping the occupancy
guard was factually wrong, in three places. reclaim_sector zeroes only
leader, so a freed sector keeps its former quote_authority and that
ex-authority's compare still passes — the removal is safe because the
write is inert (it cannot reach the free-list links, and matching walks
the active DLL only), not because the compare rejects it. The unit test
standing in for the guard was passing vacuously against a sector whose
authority had never been populated; it now reclaims a real one and bounds
the write's blast radius by byte index.

The instruction-data length still goes unchecked before the 160-byte
copy, which the reference build rejects at deserialization. Left that way
deliberately — the destination is a fixed window inside the caller's own
bounds-checked sector, so malformed data cannot corrupt anything beyond
the caller's own ladder — but the divergence and its silent 133-164 byte
window are now named in the carve-out rather than left implicit, and
tracked for a cheap follow-up guard.

Also: an end-to-end test for the newly-reachable path (write an over-cap
ladder through the instruction, take against it, watch the side drop out
of the book), a wire-encoding assertion in place of one that compared two
compile-time constants, byte-exactness asserted against the stored region
rather than inferred from which bytes changed, and the accepted-case test
names moved onto the sibling's convention.

+++ programs/dropset/src/state/market/reference_price.rs
@@ -129,7 +129,7 @@ mod tests {
         let before = data.clone();
         // `SECTORS` is one past the last live sector.
         assert_eq!(
-            stamp_reference_price(&mut data, SECTORS as u32, 1, 1, &AUTH),
+            stamp_reference_price(&mut data, SECTORS, 1, 1, &AUTH),
             Err(err::INVALID_SECTOR_INDEX)
         );
         // The null-sector sentinel is the worst case.
@@ -146,9 +146,9 @@ mod tests {
         // post-external-resize edge `Slab::effective_len` guards). The
         // capacity leg must still reject, matching `min(len, capacity)`.
         let mut data = market_buf(0);
-        write_u32(&mut data, LEN_OFF, (SECTORS as u32) + 2);
+        write_u32(&mut data, LEN_OFF, SECTORS + 2);
         assert_eq!(
-            stamp_reference_price(&mut data, SECTORS as u32, 1, 1, &AUTH),
+            stamp_reference_price(&mut data, SECTORS, 1, 1, &AUTH),
             Err(err::INVALID_SECTOR_INDEX)
         );
     }
